Resturcturing project (from guide with nginx+docker)
