--- .env.db.prod
@@ -1,4 +1,0 @@
-POSTGRES_USER=arima
-POSTGRES_PASSWORD=arima
-PGADMIN_EMAIL=arima@email
-PGADMIN_PASSWORD=arima-password
