+++ .env.dev
@@ -1,3 +1,7 @@
-FLASK_APP=app/__init__.py
+FLASK_APP=project/app.py
 FLASK_ENV=development
-APP_FOLDER=/usr/src/app
+DATABASE_URL=postgresql://arima:arima@db:5432/movies
+SQL_HOST=db
+SQL_PORT=5432
+DATABASE=postgres
+APP_FOLDER=/usr/src/main
