+++ .env.prod
@@ -1,3 +1,7 @@
-FLASK_APP=app/__init__.py
-FLASK_ENV=development
-APP_FOLDER=/usr/src/app
+FLASK_APP=project/app.py
+FLASK_ENV=production
+DATABASE_URL=postgresql://arima:arima@db:5432/movies
+SQL_HOST=db
+SQL_PORT=5432
+DATABASE=postgres
+APP_FOLDER=/home/app/web
