+++ .env.prod.db
@@ -1,6 +1,3 @@
 POSTGRES_USER=arima
 POSTGRES_PASSWORD=arima
 POSTGRES_DB=movies
-SQL_HOST=db
-SQL_PORT=5432
-POSTGRES_URL
