--- services/backend/requirements.txt
@@ -1,38 +1,0 @@
-alembic==1.6.5
-aniso8601==9.0.1
-astroid==2.5
-attrs==21.2.0
-click==7.1.2
-Flask==1.1.4
-Flask-Login==0.5.0
-Flask-Migrate==3.0.1
-flask-restplus==0.13.0
-flask-restx==0.4.0
-Flask-SQLAlchemy==2.5.1
-greenlet==1.1.0
-gunicorn==20.1.0
-iniconfig==1.1.1
-isort==5.9.1
-itsdangerous==1.1.0
-Jinja2==2.11.3
-jsonschema==3.2.0
-lazy-object-proxy==1.6.0
-Mako==1.1.4
-MarkupSafe==2.0.1
-mccabe==0.6.1
-packaging==20.9
-pluggy==0.13.1
-psycopg2==2.9.1
-py==1.10.0
-pylint==2.6.0
-pyparsing==2.4.7
-pyrsistent==0.18.0
-pytest==6.2.4
-python-dateutil==2.8.1
-python-editor==1.0.4
-pytz==2021.1
-six==1.16.0
-SQLAlchemy==1.4.20
-toml==0.10.2
-Werkzeug==1.0.1
-wrapt==1.12.1
