+++ services/nginx/Dockerfile
@@ -1,5 +1,4 @@
 FROM nginx:1.19-alpine
 
 RUN rm /etc/nginx/conf.d/default.conf
-
 COPY nginx.conf /etc/nginx/conf.d
