+++ services/web/project/config.py
@@ -1,0 +1,11 @@
+import os
+
+
+basedir = os.path.abspath(os.path.dirname(__file__))
+
+
+class Config(object):
+    SQLALCHEMY_DATABASE_URI = os.getenv("DATABASE_URL", 'postgresql://')
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
+    STATIC_FOLDER = f"{os.getenv('APP_FOLDER')}/project/static"
+    MEDIA_FOLDER = f"{os.getenv('APP_FOLDER')}/project/media"
