+++ services/web/project/db.py
@@ -1,0 +1,4 @@
+from flask_sqlalchemy import SQLAlchemy
+from sqlalchemy.dialects import INTEGER, SMALLINT, VARCHAR, TEXT, TIMESTAMP, BOOLEAN 
+
+db = SQLAlchemy()
