+++ services/web/project/static/hello.txt
@@ -1,0 +1,1 @@
+hi!
