+++ services/web/requirements.txt
@@ -1,0 +1,4 @@
+Flask==1.1.2
+Flask-SQLAlchemy==2.5.1
+gunicorn==20.1.0
+psycopg2-binary==2.8.6
